requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
apscheduler==3.10.4
aiosqlite==0.20.0
python-dotenv==1.0.1
cryptography>=42.0.0        # Ed25519 подпись для Backpack
hyperliquid-python-sdk       # SDK для Hyperliquid
curl-cffi>=0.7.0             # Chrome TLS fingerprint для Variational (обход Cloudflare)
eth-account>=0.11.0          # SIWE-авторизация для авто-обновления Variational токена
x10-python-trading-starknet  # SDK для торговли на Extended (StarkNet)
